Changed default handler delegate method
In many cases (especially on the iPad), if a new controller is
presented in response to an action of UIActionSheet by using the
clickedButtonAtIndex delegate method, the presentation fail because the
action sheet is still shown (a warning in the console confirm this). By
calling the default handlers on DidDismiss this problem is solved.